Create Terms and Conditions.md
July, 2024
Terms of  Service and  Privacy


Last updated: July 7, 2024


AGREEMENT TO TERMS

You agree that by accessing the service, you have read, understood, and agree to be bound by all of these Terms and Conditions.


SERVICE MANAGEMENT

We reserve the right, but not the obligation, to:

    (1) monitor the service for violations of these Terms and Conditions;

    (2) take appropriate legal action against anyone who, in our sole discretion, violates the law or these Terms and Conditions, including without limitation, reporting such user to law enforcement authorities;

    (3) in our sole discretion and without limitation, notice, or liability, to restrict access to the service.

    (4) monitor the proper functioning of the service.


PRIVACY POLICY

   We do NOT retain any information of your client software nor querying activity.

   We do NOT record personally identifiable information (PII) of users accessing the service, including user IP, browser information.

   We do NOT share any information of our servers activity.

   We do NOT use cookies in our website.

   We do NOT have any agreement to sell or give information to third-party actors.

TERM AND TERMINATION

These Terms and Conditions shall remain in full force and effect while you use the service.


MODIFICATIONS AND INTERRUPTIONS

We cannot guarantee the service will be available at all times. We may experience hardware, software, or other problems or need to perform maintenance related to the service, resulting in interruptions, delays, or errors.

We reserve the right to change, revise, update, suspend, discontinue, or otherwise modify the service at any time or for any reason without notice to you. You agree that we have no liability whatsoever for any loss, damage, or inconvenience caused by your inability to access or use the service during any downtime or discontinuance of the service.

Nothing in these Terms and Conditions will be construed to obligate us to maintain and support the service or to supply any corrections, updates, or releases in connection therewith.


GOVERNING LAW

These Terms and Conditions and your use of the service are governed by and construed in accordance with the laws of European Union and GDPR Privacy Policy (Data Protection Rights).